requirements.txt
discord.py>=2.4.0
apscheduler>=3.10.4
aiohttp>=3.9.1
python-dotenv>=1.0.1
psutil>=5.9.8
mss>=9.0.1
pyotp>=2.9.0
